+++ scripts/state/game_state.gd
@@ -1,6 +1,6 @@
 extends Node
 
-enum VictoryType { NONE, HERO_WIN, TROLL_WIN }
+enum VictoryType { NONE, HERO_WIN, TROLL_WIN, ESCAPE_WIN }
 enum Faction { HERO, BOSS }
 
 var selected_faction: Faction = Faction.HERO
@@ -8,6 +8,13 @@ var selected_character: String = "Warrior"
 var last_selected_hero: String = "Warrior"
 var is_game_over: bool = false
 var victory_type: VictoryType = VictoryType.NONE
+var player_gold: int = 0
+
+signal gold_changed(amount: int)
+
+func add_gold(amount: int):
+	player_gold += amount
+	gold_changed.emit(player_gold)
 
 func end_game(type: VictoryType):
 	is_game_over = true
@@ -18,3 +25,4 @@ func reset():
 	selected_character = "Warrior"
 	is_game_over = false
 	victory_type = VictoryType.NONE
+	player_gold = 0